3_Replacing Outlier with Mean /Median / Mode
This notebook demonstrated a practical method of handling outliers in a dataset by replacing them with the mean. This technique is commonly used when you have a normally distributed dataset with extreme values that distort statistical measures like the mean. By replacing outliers with the mean, we can make the data more stable and reliable for further analysis.